+++ app/controllers/home.js
@@ -1,0 +1,3 @@
+module.exports.page_home = function(application, req, res){
+    res.render('home/index');
+};
