+++ app/routes/home.js
@@ -1,6 +1,6 @@
-module.exports = function(app){
-    app.get('/',function(req,res){
-        res.render('home/index');
+module.exports = function(application){
+    application.get('/',function(req,res){
+       application.app.controllers.home.page_home(application, req, res)
     });
 };
