+++ app/routes/noticias.js
@@ -1,14 +1,12 @@
-var dbConnection = require('../../config/dbConnection');
 
 module.exports = function(application){
-
    application.get('/noticias',function(req,res){
-      var connection = dbConnection()
-      var noticiasModel = new application.app.models.noticiasModel(connection);
-
-         noticiasModel.getNoticias(function(error, result){
-         res.render('noticias/noticias',{noticias: result});
-      });
+      application.app.controllers.noticias.list_noticias(application, req, res)
+     
    });     
+
+   application.get('/noticia',function(req,res){
+     application.app.controllers.noticias.list_noticia(application, req, res)
+   });
 };
