+++ app/views/admin/form_add_noticia.ejs
@@ -16,6 +16,7 @@
 			</ul>
 		<% }%>
 
+		
 		<h1>Adicionar Notícia</h1>
 		</br>
 		<form action="/noticias/salvar" method="post">
@@ -43,7 +44,7 @@
 			<label>notícia</label>
 			
 		</br>
-			<textarea name="noticia" id="noticia" value="<%=noticias.noticia%>" cols="21" rows="10" placeholder="Descreva a Notícia"></textarea>></textarea>
+			<textarea name="noticia" id="noticia" value="<%=noticias.noticia%>" cols="21" rows="10" placeholder="Descreva a Notícia"></textarea></textarea>
 			
 		</br> 
 			<input type="submit" value="Enviar"/>
